services/ocr/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
PyYAML==6.0.1
minio==7.2.5
pypdf==4.0.2
